Added fixes to reported issues
Big change: now logs setup progress to /var/log/hermit.log and displays a summary of that at startup

Also made changes to suspend_on_idle to attempt to workaround issue where the machine suspended but it didn't think it suspended and so it shuts down the machine entirely.

Lastly put in "assumptions": A record of things have been done so next time we don't need to check next time.